Made Property Summary sortable by price and distance.
Added distance property. Right now, this is just a counter so closest property has distance of 0, second closest property has distance of 1, and so forth. Made price attribute an integer.

Now sorts results by price or distance using a sort descriptor in the fetched results controller. Use to rely on the server handling the sorting. Still passes sort information to the server for statistics.