Does a deep copy of the Criteria object when adding to the History object. Deletes History objects with 0 results.
Scenario:
A user does a search with 0 results. The History object is immediately deleted, cascading to Criteria as well. If the user goes back and loosens up his Criteria, then re-performs the search and gets results, the app will fail when trying to save the context. The reason is because it will be trying to save a Criteria object that has actually been deleted the first time.

Scenario 2:
If a user performs a search with results, then goes back and performs the search again with modified criteria and gets results, both History objects will have the same Criteria object, which is wrong.