Removed setTitle from all root views. Sets title in initWithNib in rest of views, where possible.
Some title logic depends on values passed after the init, so those are kept in viewDidLoad. Moved titles to initWithNib so when the views are restored, the page after that view will display the proper title in the back button instead of "Back" since the title will have been set.