Fixed new bug where Google Maps now requires and API key and added delay to geocoding requests.
This commit is meant to be used for Find an Apartment 2.1 and Home Finder 2.0.1 (2.0.1 because submitted an update to Home Finder earlier with that version and Apple does not allow version changes for updated after submitted for review).

For both apps, the update is quadruple the number of cities available and fix the key issue with Google Maps and add a delay to Google Map request to prevent a too many requests error.